Add initial remote data structure to fetch data from API
