--- app/src/main/java/com/mypackage/sportsevents/data/mapper/EventMapper.kt
@@ -1,12 +1,0 @@
-package com.mypackage.sportsevents.data.mapper
-
-import com.mypackage.sportsevents.data.EventDto
-import com.mypackage.sportsevents.domain.Event
-
-fun EventDto.toDomain(): Event {
-    return Event(
-        id = id,
-        name = name,
-        timestamp = timestamp
-    )
-}
