+++ app/src/main/java/com/mypackage/sportsevents/data/remote/api/RetrofitInstance.kt
@@ -1,0 +1,17 @@
+package com.mypackage.sportsevents.data.remote.api
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitInstance {
+
+    private const val BASE_URL = "https://ios-kaizen.github.io/"
+
+    val api: SportsApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(SportsApi::class.java)
+    }
+}
