+++ app/src/main/java/com/mypackage/sportsevents/data/remote/api/SportsApi.kt
@@ -1,0 +1,9 @@
+package com.mypackage.sportsevents.data.remote.api
+
+import com.mypackage.sportsevents.data.remote.dto.SportDto
+import retrofit2.http.GET
+
+interface SportsApi {
+    @GET("MockSports/sports.json")
+    suspend fun getSports(): List<SportDto>
+}
