+++ app/src/main/java/com/mypackage/sportsevents/data/repository/SportsRepositoryImpl.kt
@@ -1,0 +1,14 @@
+package com.mypackage.sportsevents.data.repository
+
+import com.mypackage.sportsevents.data.remote.api.SportsApi
+import com.mypackage.sportsevents.data.remote.mapper.toDomain
+import com.mypackage.sportsevents.domain.model.Sport
+import com.mypackage.sportsevents.domain.repository.SportsRepository
+
+class SportsRepositoryImpl(
+    private val api: SportsApi
+) : SportsRepository {
+    override suspend fun getAllSports(): List<Sport> {
+        return api.getSports().map { it.toDomain() }
+    }
+}
