+++ app/src/main/java/com/mypackage/sportsevents/domain/model/Event.kt
@@ -1,4 +1,4 @@
-package com.mypackage.sportsevents.domain
+package com.mypackage.sportsevents.domain.model
 
 data class Event(
     val id: String,
