+++ app/src/main/java/com/mypackage/sportsevents/domain/model/Sport.kt
@@ -1,4 +1,4 @@
-package com.mypackage.sportsevents.domain
+package com.mypackage.sportsevents.domain.model
 data class Sport(
     val id: String,
     val name: String,
