+++ app/src/main/java/com/mypackage/sportsevents/domain/repository/SportsRepository.kt
@@ -1,0 +1,7 @@
+package com.mypackage.sportsevents.domain.repository
+
+import com.mypackage.sportsevents.domain.model.Sport
+
+interface SportsRepository {
+    suspend fun getAllSports(): List<Sport>
+}
